MAC binary changes

+++ hack/test-ocp-etcd-backup-restore-plugin-locally.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 
 # This script verifies that a ocp-etcd-backup-restore build can be installed to a system using
+
 # krew local testing method
 
 set -euo pipefail
@@ -10,7 +11,8 @@ set -euo pipefail
 SRC_ROOT="$(cd "$(dirname "${BASH_SOURCE[0]}")" && cd .. && pwd)"
 cd "$SRC_ROOT"
 
-build_dir="build"
+ocp_dir="ocp_etcd_backup_restore"
+build_dir="build/$ocp_dir"
 
 ocp_etcd_backup_restore_manifest="${build_dir}/ocp-etcd-backup-restore.yaml"
 if [[ ! -f "${ocp_etcd_backup_restore_manifest}" ]]; then
@@ -18,7 +20,8 @@ if [[ ! -f "${ocp_etcd_backup_restore_manifest}" ]]; then
   exit 1
 fi
 
-ocp_etcd_backup_restore_archive="${build_dir}/ocp-etcd-backup-restore.tar.gz"
+ocp_etcd_backup_restore_tar="ocp-etcd-backup-restore-Linux.tar.gz"
+ocp_etcd_backup_restore_archive="${build_dir}/${ocp_etcd_backup_restore_tar}"
 if [[ ! -f "${ocp_etcd_backup_restore_archive}" ]]; then
   echo >&2 "Could not find archive ${ocp_etcd_backup_restore_archive}."
   exit 1
@@ -27,6 +30,13 @@ fi
 # test for linux OS
 kubectl krew install --manifest=$ocp_etcd_backup_restore_manifest --archive=$ocp_etcd_backup_restore_archive
 kubectl krew uninstall ocp-etcd-backup-restore
+
+ocp_etcd_backup_restore_tar="ocp-etcd-backup-restore-macOS.tar.gz"
+ocp_etcd_backup_restore_archive="${build_dir}/${ocp_etcd_backup_restore_tar}"
+if [[ ! -f "${ocp_etcd_backup_restore_archive}" ]]; then
+  echo >&2 "Could not find archive ${ocp_etcd_backup_restore_archive}."
+  exit 1
+fi
 
 # test for darwin OS
 KREW_OS=darwin KREW_ARCH=amd64 kubectl krew install --manifest=$ocp_etcd_backup_restore_manifest --archive="$ocp_etcd_backup_restore_archive"
